pyproject.toml
[project]
name = "sillypaste"
version = "0.0.1"

[tool.setuptools]
packages = ["sillypaste"]

[tool.black]
skip-string-normalization = true
line-length = 79

[tool.poetry]
name = "sillypaste"
version = "0.0.1"
description = ""
readme = "README.md"
packages = [{include = "sillypaste"}]
authors = ["Winston Weinert <hello@winny.tech>"]

[tool.poetry.scripts]
sillypaste = "sillypaste.manage:main"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-watchman = "^1.3.0"
django-lazysignup = {git = "https://github.com/winny-/django-lazysignup.git", rev = "ec753db87e21df4b4489c0ef31a86595b9277605"}
django-canonical-domain = "^0.9.2"
org-python = "^0.3.2"
markdown = "^3.3.4"
pygments = "^2.8.1"
dj-database-url = "^2.0.0"
gunicorn = "^20.1.0"
Django.version = "^4.1.6"
django-heroku.version = "^0.3.1"
psycopg2.version = "^2.8.6"
python-dotenv.version = "^0.19.1"
django-livereload-server.version = "^0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
